+++ src/core/net/peer.h
@@ -5,15 +5,18 @@
 #include <sockpp/tcp_acceptor.h>
 #include <sockpp/tcp_connector.h>
 #include "../bytebuff.h"
+#include "../components/role.h"
 
 class peer {
 private:
+    role type = role::NODE;
     vec<str> peers;
     sockpp::tcp_acceptor acceptor;
     sockpp::tcp_connector connector;
 public:
-    peer() = default;
+    peer();
     explicit peer(u16 port);
+    explicit peer(role);
     ~peer();
 
     void addPeer(const str& addr);
@@ -22,6 +25,7 @@ public:
     void broadcast(const bytebuff& b);
 
     u32 peerCount();
+    role getRole();
 
     sockpp::socket accept();
 
@@ -29,6 +33,7 @@ public:
     bytebuff recv();
 
     void open(u16 port);
+    void open(role);
     void close();
 };