+++ src/core/utils.h
@@ -4,16 +4,11 @@
 #include "constants.h"
 #include <chrono>
 
-template<class I, class O>
-vec<O> mapTo(const vec<I>& inputs, O (f)(const I&)) {
-    auto temp = vec<O>(inputs.size());
-    for(i32 i = 0;i < inputs.size();i++){
-        temp.push_back(f(inputs[i]));
-    }
-    return temp;
-}
+template<u64 size>
+str to_string(const arr<u64,size>&);
 
-u64 timestamp();
+template<u64 size>
+arr<u64,size> to_arr(const str&);
 
 str to_string(const hash256&);
 str to_string(const hash384&);
@@ -26,5 +21,7 @@ hash512 to_hash512(const str&);
 std::ostream& operator << (std::ostream& cout, const hash256& hash);
 std::ostream& operator << (std::ostream& cout, const hash384& hash);
 std::ostream& operator << (std::ostream& cout, const hash512& hash);
+
+u64 timestamp();
 
 #endif //FCOIN_UTILS_H